src/python/requirements.txt
selenium
bs4
python-dotenv
supabase